refactored scientificName and Locality generation into catalogOfLife.py and locality.py respectively
